Fix build dependencies on certain arches
https://bugzilla.redhat.com/show_bug.cgi?id=1859515

Note this might not help on arches like s390x, where there's still quite
a few dependencies we won't have. This might also mean if the SRPM gets
generated on s390x `dnf builddep` might not pull in all dependencies,
but that's better then having it fail most of the time on all arches.