Prevent partial updates (rhbz#2193135)
This is another attempt to prevent partial updates, where important mesa
subpackages get desynchronized with mesa-dri-drivers, and things crash because
of version mismatch.

According to Michel Dänzer [1], libglapi and libgbm should always match in
version to dri-drivers. So the conditional requirement was added there, and it
will help guard dependants (e.g. libGL, libEGL) versions in the future.

This is the Rawhide-version of PR 24 [2].

[1] https://bugzilla.redhat.com/show_bug.cgi?id=2187726#c4
[2] https://src.fedoraproject.org/rpms/mesa/pull-request/24