Disable NVK for now in ELN builds
While it is possible that we may eventually want to enable this for
RHEL, in its current form it is pulling in OpenCL dependencies (e.g.
libclc, SPIRV) and uses Fedora rust-packaging and crates.  Because it
does not use the usual %cargo_* macros, it is not clear how to go about
vendoring the rust dependencies for ELN and RHEL builds.