Fix up some directory ownerships
The vulkan drivers package ships some files in a directory
owned by the filesystem package, and there's another dir
that has crossover so move it to the filesystem package.

Move the vdpau to it's subpackage as it's only used by that
package so it doesn't need to be in filesystem.